+++ lib/credit_card_validations/error.rb
@@ -1,0 +1,4 @@
+module CreditCardValidations
+  class Error < StandardError
+  end
+end