+++ lib/credit_card_validations/plugins/diners_us.rb
@@ -1,0 +1,6 @@
+CreditCardValidations.add_brand(
+  :diners_us,
+  { length: 16, prefixes: %w(54 55) },
+  brand_name: 'Diners Club US'
+)
+