+++ lib/credit_card_validations/plugins/en_route.rb
@@ -1,0 +1,5 @@
+CreditCardValidations.add_brand(
+  :en_route,
+  { length: 15, prefixes: %w(2014 2149) },
+  skip_luhn: true
+)