+++ lib/credit_card_validations/plugins/laser.rb
@@ -1,0 +1,4 @@
+CreditCardValidations.add_brand(
+  :laser,
+  length: [16, 17, 18, 19], prefixes: %w(6304 6706 6771)
+)