+++ spec/active_model_spec.rb
@@ -1,13 +1,13 @@
 require_relative 'test_helper'
 
-describe "ActiveModel Validator" do
+describe 'ActiveModel Validator' do
 
   let(:model) { CreditCard.new }
 
-  describe "Proc support" do
-    it "should be valid if brands from proc valid" do
+  describe 'Proc support' do
+    it 'should be valid if brands from proc valid' do
        card = model.dup
-       card.card_type =  "Master Card"
+       card.card_type = 'Master Card'
        card.number6 =  CreditCardValidations::Factory.random(:visa)
        card.valid?.must_equal false
        card.number6 =  CreditCardValidations::Factory.random(:mastercard)
@@ -15,8 +15,8 @@ describe "ActiveModel Validator" do
     end
   end
 
-  describe "Any Brand" do
-    it "should be valid for all prepared valid numbers" do
+  describe 'Any Brand' do
+    it 'should be valid for all prepared valid numbers' do
       VALID_NUMBERS.each do |_, numbers|
         numbers.each do |number|
           card = model
@@ -28,8 +28,8 @@ describe "ActiveModel Validator" do
     end
   end
 
-  describe "Except Amex and Maestro brand" do
-    it "should reject all other valid numbers" do
+  describe 'Except Amex and Maestro brand' do
+    it 'should reject all other valid numbers' do
       VALID_NUMBERS.except(:amex, :maestro).each do |_, numbers|
         card = model
         card.number = numbers.first
@@ -38,7 +38,7 @@ describe "ActiveModel Validator" do
       end
     end
 
-    it "should accept using except options" do
+    it 'should accept using except options' do
       VALID_NUMBERS.except(:amex, :maestro).each do |_, numbers|
         card = model
         card.number3 = numbers.first
@@ -47,8 +47,8 @@ describe "ActiveModel Validator" do
     end
   end
 
-  describe "Only Amex and Mestro brands" do
-    it "should accept amex and maestro brand if valid" do
+  describe 'Only Amex and Mestro brands' do
+    it 'should accept amex and maestro brand if valid' do
       VALID_NUMBERS.slice(:amex, :maestro).each do |_, numbers|
         card = model
         card.number = numbers.first
@@ -58,12 +58,12 @@ describe "ActiveModel Validator" do
     end
   end
 
-  describe "Custom error message"  do
-    it "should allow custom message" do
+  describe 'Custom error message' do
+    it 'should allow custom message' do
       card = model
       card.number7 =  'wrong'
       card.valid?.must_equal false
-      card.errors[:number7].must_equal ["Custom message"]
+      card.errors[:number7].must_equal ['Custom message']
     end
   end